FP-95, FP-104, FP-107, FP-110, FP-112, FP-114

+++ server/routes/api/meals/index.js
@@ -1,16 +1,22 @@
 const   meals           = require('express').Router()
     ,   mealsDB         = require(process.env.FOOD_HOME + 'modules/db/meals')
-    ,   signupsDB         = require(process.env.FOOD_HOME + 'modules/db/signups')
+    ,   signupsDB       = require(process.env.FOOD_HOME + 'modules/db/signups')
     ,   paymentDB       = require(process.env.FOOD_HOME + 'modules/db/payment')
     ,   image           = require(process.env.FOOD_HOME + 'middleware/singleImage')
     ,   error           = require(process.env.FOOD_HOME + 'modules/error')
     ,   scheduler       = require(process.env.FOOD_HOME + 'modules/scheduler')
     ,   mail            = require(process.env.FOOD_HOME + 'modules/mailer')
+    ,   caches          = require(process.env.FOOD_HOME + 'modules/cache')
+    ,   jwt             = require(process.env.FOOD_HOME + 'modules/auth/jwt')
     ,   notification    = require(process.env.FOOD_HOME + 'modules/notification')
     ,   fs              = require('fs')
     ,   log             = require(process.env.FOOD_HOME + 'modules/log');
 
-meals.post('/:id/lock', error.router.validate('params', {
+let cache = caches.getCache('meals'),
+    signupCache = caches.getCache('signups'),
+    updateCache = caches.getCache('update');
+
+meals.post('/:id/lock', jwt.requireAuthentication, error.router.validate('params', {
     id: /^[0-9]{1,9}$/
 }), error.router.validate('body', {
     prices: 'array'
@@ -32,7 +38,20 @@ meals.post('/:id/lock', error.router.validate('params', {
         log(4, 'PriceArray not valid.');
         return res.status(400).send({msg: 'Options not valid.', type: 'Invalid_Request', data: [JSON.stringify(errorElem)]});
     }
-    paymentDB.setPrices(req.body.prices)
+    cache.delete(req.params.id);
+    cache.delete('allMeals');
+    signupCache.deleteAll();
+    updateCache.deleteAll();
+
+    mealsDB.getMealById(req.params.id)
+        .then(meal => {
+            if (meal.creatorId == req.user.id) {
+                return paymentDB.setPrices(req.body.prices);
+            } else {
+                log(4, `User ${req.user.id} tried to lock meal ${req.params.id} without being the creator.`);
+                return Promise.reject({status: 403, type: 'FORBIDDEN'});
+            }
+        })
         .then(() => paymentDB.lockMealPrices(req.params.id))
         .then(() => paymentDB.getEligibleSignups(req.params.id))
         .then((eligibleSignups) => Promise.all(
@@ -51,7 +70,9 @@ meals.post('/:id/lock', error.router.validate('params', {
         .catch(error.router.internalError(res));
 });
 
-meals.post('/prices', error.router.validate('body', {
+meals.post('/:id/prices', jwt.requireAuthentication, error.router.validate('params', {
+    id: /^[0-9]{1,9}$/
+}), error.router.validate('body', {
     prices: 'array'
 }), (req, res) => {
     let errorElem,
@@ -72,34 +93,42 @@ meals.post('/prices', error.router.validate('body', {
         return res.status(400).send({msg: 'Options not valid.', type: 'Invalid_Request', data: [JSON.stringify(errorElem)]});
     }
 
-    paymentDB.setPrices(req.body.prices)
+    cache.delete(req.params.id);
+    cache.delete('allMeals');
+    updateCache.deleteAll();
+
+    mealsDB.getMealById(req.params.id)
+        .then(meal => {
+            if (meal.creatorId == req.user.id) {
+                return paymentDB.setPrices(req.body.prices);
+            } else {
+                log(4, `User ${req.user.id} tried to set prices for meal ${req.params.id} without being the creator.`);
+                return Promise.reject({status: 403, type: 'FORBIDDEN'});
+            }
+        })
         .then(result => {
             res.status(200).send(result);
         })
         .catch(error.router.internalError(res));
 });
 
-meals.get('/:id', error.router.validate('params', {
-    id: /^[0-9]{1,9}$/
-}), (req, res) => {
-    mealsDB.getMealById(req.params.id).then((meals) => {
-        res.status(200).send(meals);
-    })
-    .catch(error.router.internalError(res));
-});
-
 meals.get('/', (req, res) => {
-    mealsDB.getAllMeals().then((meals) => {
-        res.status(200).send(meals);
-    })
-    .catch(error.router.internalError(res));
+    let meal = cache.get('allMeals');
+    if (meal) {
+        res.status(200).send(meal);
+    } else {
+        mealsDB.getAllMeals().then((meals) => {
+            cache.put('allMeals', meals);
+            res.status(200).send(meals);
+        })
+        .catch(error.router.internalError(res));
+    }
 });
 
-meals.put('/:id', image.single('imageData'), error.router.validate('params', {
+meals.put('/:id', image.single('imageData'), jwt.requireAuthentication, error.router.validate('params', {
     id: /^[0-9]{1,9}$/
 }), error.router.validate('body', {
     name: /^[ÄÜÖäöüA-Za-z0-9.\-,\s]{2,70}$/,
-    creator: /^[ÄÜÖäöüA-Za-z0-9.\-,\s]{0,70}$/,
     description: /^[^"%;]*$/,
     time: /^[0-9]{1,15}$/,
     deadline: /^[0-9]{0,15}$/,
@@ -138,9 +167,21 @@ meals.put('/:id', image.single('imageData'), error.router.validate('params', {
         delete mealData.image;
     }
 
-    mealsDB.setMealById(req.params.id, mealData)
-        .then(id => mealsDB.getMealById(id))
-        .then((meal) => {
+    cache.delete(req.params.id);
+    cache.delete('allMeals');
+    updateCache.deleteAll();
+
+    mealsDB.getMealById(req.params.id)
+    .then(meal => {
+        if (meal.creatorId == req.user.id) {
+            return mealsDB.setMealById(req.params.id, mealData);
+        } else {
+            log(4, `User ${req.user.id} tried to update meal ${req.params.id} without being the creator.`);
+            return Promise.reject({status: 403, type: 'FORBIDDEN'});
+        }
+    })
+    .then(id => mealsDB.getMealById(id))
+    .then((meal) => {
         scheduler.rescheduleMeal(meal);
         if (req.file) {
             fs.readdir(process.env.FOOD_CLIENT + '/images/meals/', function (err, files) {
@@ -165,10 +206,24 @@ meals.put('/:id', image.single('imageData'), error.router.validate('params', {
     .catch(error.router.internalError(res));
 });
 
-meals.delete('/:id', error.router.validate('params', {
+meals.delete('/:id', jwt.requireAuthentication, error.router.validate('params', {
     id: /^[0-9]{1,9}$/
 }), (req, res) => {
-    mealsDB.deleteMealById(req.params.id)
+
+    cache.delete(req.params.id);
+    cache.delete('allMeals');
+    signupCache.deleteAll();
+    updateCache.deleteAll();
+
+    mealsDB.getMealById(req.params.id)
+    .then(meal => {
+        if (meal.creatorId == req.user.id) {
+            return mealsDB.deleteMealById(req.params.id);
+        } else {
+            log(4, `User ${req.user.id} tried to delete meal ${req.params.id} without being the creator.`);
+            return Promise.reject({status: 403, type: 'FORBIDDEN'});
+        }
+    })
     .then(() => signupsDB.deleteSignupsByMeal(req.params.id))
     .then((data) => {
         scheduler.cancelMeal(req.params.id);
@@ -191,7 +246,7 @@ meals.delete('/:id', error.router.validate('params', {
     .catch(error.router.internalError(res));
 });
 
-meals.post('/', image.single('imageData'), error.router.validate('body', {
+meals.post('/', image.single('imageData'), jwt.requireAuthentication, error.router.validate('body', {
     name: /^[ÄÜÖäöüA-Za-z0-9.\-,\s]{2,70}$/,
     creator: /^[ÄÜÖäöüA-Za-z0-9.\-,\s]{0,70}$/,
     description: /^[^"%;]*$/,
@@ -236,6 +291,8 @@ meals.post('/', image.single('imageData'), error.router.validate('body', {
     mealsDB.createMeal(mealData)
         .then(id => mealsDB.getMealById(id))
         .then(meal => {
+            cache.delete('allMeals');
+            updateCache.deleteAll();
             mail.sendCreationNotice(meal);
             scheduler.scheduleMeal(meal);
             notification.sendCreationNotice(meal);
@@ -254,7 +311,7 @@ meals.post('/', image.single('imageData'), error.router.validate('body', {
 });
 
 
-meals.post('/:id/mail', error.router.validate('params', {
+meals.post('/:id/mail', jwt.requireAdmin, error.router.validate('params', {
     id: /^[0-9]{1,9}$/
 }), (req, res) => {
     mealsDB.getMealById(req.params.id).then((meals) => {
